fix: client cannot import server-only file
