+++ front/lib/api/assistant/global_agents.ts
@@ -4,19 +4,18 @@ import { promisify } from "util";
 
 const readFileAsync = promisify(fs.readFile);
 
+import {
+  CLAUDE_DEFAULT_MODEL_CONFIG,
+  CLAUDE_INSTANT_DEFAULT_MODEL_CONFIG,
+  GPT_3_5_TURBO_DEFAULT_MODEL_CONFIG,
+  GPT_4_DEFAULT_MODEL_CONFIG,
+} from "@app/lib/api/assistant/supported_models";
 import { Authenticator, prodAPICredentialsForOwner } from "@app/lib/auth";
 import { ConnectorProvider } from "@app/lib/connectors_api";
 import { DustAPI } from "@app/lib/dust_api";
 import logger from "@app/logger/logger";
 import { AgentConfigurationType } from "@app/types/assistant/agent";
 import { UserType } from "@app/types/user";
-
-import {
-  CLAUDE_DEFAULT_MODEL_CONFIG,
-  CLAUDE_INSTANT_DEFAULT_MODEL_CONFIG,
-  GPT_3_5_TURBO_DEFAULT_MODEL_CONFIG,
-  GPT_4_DEFAULT_MODEL_CONFIG,
-} from "./generation";
 
 class HelperAssistantPrompt {
   private static instance: HelperAssistantPrompt;
